fix wrong number comparison due to missing cast, e.g. 10<8

+++ 2022/13/ds1-intuition.sql
@@ -46,7 +46,7 @@ worker(i, idx, parent, id, ids, fst, snd, rec, res) AS (
     UNION ALL
     (
     -- both are numbers: compare the numbers
-    SELECT * REPLACE(i+1 AS i, false AS rec, comp(fst, snd) AS res)
+    SELECT * REPLACE(i+1 AS i, false AS rec, comp(fst::Int, snd::Int) AS res)
     FROM worker
     WHERE rec AND NOT fst ^@ '[' AND NOT snd ^@ '['
